Delete groups

+++ src/components/CardSpecialList.vue
@@ -1,6 +1,16 @@
 <template>
   <div>
     <div class="card m-3" style="width: 18rem">
+      <div class="controll-icon-container">
+        <font-awesome-icon
+          class="group-card-controll-icon"
+          :style="{ color: 'red' }"
+          icon="times-circle"
+          @click.stop="deleteGroup"
+          size="lg"
+        />
+      </div>
+
       <div class="image-zoom">
         <img class="card-img-top" :src="group.img" alt="Card image cap" />
       </div>
@@ -17,13 +27,6 @@
             {{ task.description }}
           </div>
         </div>
-        <font-awesome-icon
-          class="table-action-icon"
-          :style="{ color: 'red' }"
-          icon="trash-alt"
-          @click.stop="deleteGroup"
-          size="lg"
-        />
       </div>
     </div>
   </div>
@@ -35,18 +38,28 @@ export default {
       type: Object,
     },
   },
+  data: () => {
+    return {
+      is_delete_btn_active: true,
+    };
+  },
   methods: {
     async deleteGroup() {
+      if (!this.is_delete_btn_active) {
+        return;
+      }
+      this.is_delete_btn_active = false;
       let status = await this.$store.dispatch('deleteGroupById', this.group);
       this.showToastr(status);
+      this.is_delete_btn_active = true;
     },
     showToastr(status) {
-      switch (status) {
-        case 'OK':
-          return this.$toastr.success('Group deleted successfully');
-        default:
-          return this.$toastr.error('Something went wrong', 'Oooopss..');
-      }
+      let map = {
+        OK: () => this.$toastr.success('Group deleted successfully'),
+        'Not Found': () =>
+          this.$toastr.error('Something went wrong', 'Oooopss..'),
+      };
+      map[status]();
     },
   },
   computed: {
